File cleanup

+++ app/database/delete_post.php
@@ -5,7 +5,11 @@ require __DIR__.'/../autoload.php';
 if (isset($_POST['id'])) {
     $post_id = filter_var($_POST['id'], FILTER_SANITIZE_NUMBER_INT);
 
-    deletePost($pdo, $post_id);
+    $post = getPost($pdo, $post_id);
+
+    if ($_SESSION['user']['id'] === $post['user_id']) {
+        deletePost($pdo, $post_id);
+    }
 
     redirect('/');
 }
